requirements.txt
psutil==7.0.0
python-telegram-bot==21.10